Fix compiler some warnings
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@2378 610dfd11-2700-0410-9b68-9ae28eac5983

+++ include/H3D/FrameBufferTextureGenerator.h
@@ -567,7 +567,7 @@ namespace H3D {
     /// \param style Options extracted from color_buffer_storage about how the base will be used to 
     /// initialize the color buffer storage. 
     /// \param index The extracted index number of the color buffer attachment in the fbo
-    bool parseColorBufferStorage( std::string color_buffer_storage, std::string& style, size_t& index );
+    bool parseColorBufferStorage( std::string color_buffer_storage, std::string& style, int& index );
 
     /// Function which only clear the specified color buffer
     /// \param src  The source FBO of which the clear will do
